Move extracting message logic to mixin
